changed int to float

+++ IWA6/scripts.js
@@ -20,7 +20,7 @@ const minuteOfDay = 0;
 // Only change below this line 
 
 if (hourOfDay === 0 && minuteOfDay === 0) {
-	const taxAsDecimal = parseInt(tax) / 100
+	const taxAsDecimal = parseFloat(tax) / 100
     const startingAfterTax = salary - (salary * taxAsDecimal)
 	const balance = startingAfterTax - transport - food - rent
     console.log('R', balance.toFixed(2))
